% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{allBurdenStats}
\alias{allBurdenStats}
\title{Calculate all burden statistics simultaneously}
\usage{
allBurdenStats(ccdata, ccstatus, esm_K, LLc_maf, LLc_maf_control = TRUE,
  normalize_calpha = FALSE, simplecount_calpha = FALSE)
}
\arguments{
\item{ccdata}{A matrix of markers (columns) and individuals (rows).  Data are coded as the number of copies of the minor allele.}

\item{ccstatus}{A vector of binary phenotype labels.  0 = control, 1 = case.}

\item{esm_K}{The number of markers to use in the calculation of ESM_K}

\item{LLc_maf}{For Li and Leal's statistic, only consider variants whose minor allele frequencies are <= maf}

\item{LLc_maf_control}{For Li and Leal's statistic, calculate MAF from controls only if TRUE, otherwise from entire sample}

\item{normalize_calpha}{If TRUE, return T/sqrt(Z), otherwise return T.}

\item{simplecount_calpha}{see Details}
}
\value{
A list of values for all burden statistics
}
\description{
Calculate all burden statistics simultaneously
}
\details{
When simplecount_alpha = FALSE, heterozygous and homozygous genotypes are treated as different numbers of observations
of the mutation.  In other wordes, simplecounts = FALSE is equivalent to colSums( ccdata[status==1,] ).  When simplecounts=TRUE,
all nonzero genotype values are treated as the value 1, equivalent to  apply(data[status==1,], 2, function(x) sum(x>0, na.rm=TRUE)).
The latter method is used by the R package AssotesteR.
}
\references{
Li, B., & Leal, S. (2008). Methods for detecting associations with rare variants for common diseases: application to analysis of sequence data. The American Journal of Human Genetics, 83(3), 311-321.

Neale, B. M., Rivas, M. A., Voight, B. F., Altshuler, D., Devlin, B., Orho-Melander, M., et al. (2011). Testing for an Unusual Distribution of Rare Variants. PLoS Genetics, 7(3), e1001322. doi:10.1371/journal.pgen.1001322

Madsen, B. E., & Browning, S. R. (2009). A groupwise association test for rare mutations using a weighted sum statistic. PLoS Genetics, 5(2), e1000384. doi:10.1371/journal.pgen.1000384

Thornton, K. R., Foran, A. J., & Long, A. D. (2013). Properties and Modeling of GWAS when Complex Disease Risk Is Due to Non-Complementing, Deleterious Mutations in Genes of Large Effect. PLoS Genetics, 9(2), e1003258. doi:10.1371/journal.pgen.1003258
}

