% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{buRden}
\alias{buRden}
\alias{buRden-package}
\title{buRden.}
\description{
An Rcpp package for performing "burden" tests of associations due to rare variants by permutation.
}
\author{
Kevin R. Thornton \email{krthornt@uci.edu}
}
\references{
\url{https://github.com/molpopgen/buRden}
}

