% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccpcgc.R
\name{calc.Gij}
\alias{calc.Gij}
\title{Equation 5 from www.pnas.org/cgi/doi/10.1073/pnas.1419064111}
\usage{
calc.Gij(genotypes)
}
\arguments{
\item{genotypes}{An individuals x mutation matrix, coded 0,1,2 = number of copies of minor alleles}
}
\description{
Equation 5 from www.pnas.org/cgi/doi/10.1073/pnas.1419064111
}
\note{
copied from Golan's original code at https://sites.google.com/site/davidgolanshomepage/software/pcgc
}
\references{
Golan et al. (2014) Measuring missing heritability: Inferring the contribution of common variants. www.pnas.org/cgi/doi/10.1073/pnas.1419064111
}

