% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cAlpha_perm}
\alias{cAlpha_perm}
\title{Permutation distribution of the c-alpha statistic}
\usage{
cAlpha_perm(data, status, nperms)
}
\arguments{
\item{data}{A matrix of markers (columns) and individuals (rows).  Data are coded as the number of copies of the minor allele.}

\item{status}{A vector of binary phenotype labels.  0 = control, 1 = case.}

\item{nperms}{The number of permutations to perform}
}
\value{
The distribution of the test statistic after nperms swapping of case/control labels
}
\description{
Permutation distribution of the c-alpha statistic
}
\examples{
data(rec.ccdata)
status = c( rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases) )
#get minor allele freqs in
rec.ccdata.MAFS = colSums( rec.ccdata$genos[which(status==0),] )/(2*rec.ccdata$ncontrols)
rec.ccdata.calpha = cAlpha(rec.ccdata$genos[,which(rec.ccdata.MAFS <= 0.05)],status)
rec.ccdata.calpha.permdist = cAlpha_perm(rec.ccdata$genos[,which(rec.ccdata.MAFS <= 0.05)],status)
}

