% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ccpvals}
\alias{ccpvals}
\title{Obtain single-marker p-values for case/control data}
\usage{
ccpvals(genos, status, model = "additive")
}
\arguments{
\item{genos}{A matrix of genotypes, coded as number of copies of minor allele}

\item{status}{An array of 0 = control, 1 = case.  length(status) must equal nrow(x)}

\item{model}{One of "additive","recessive", or "dominant"}
}
\value{
An array of p-values of logistic regressions of case/control status onto genotype.  The order of the p-values corresponds to the column order in x.
}
\description{
Obtain single-marker p-values for case/control data
}
\examples{
data(rec.ccdata)
#The function works on the genotype matrix
#We assign discrete phenotypes based on number of controls & cases in ccblock.dsdata$genos:
pvals.additive = ccpvals(rec.ccdata$genos,c(rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases)))
pvals.recessive = ccpvals(rec.ccdata$genos,c(rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases)),"recessive")
pvals.dominant = ccpvals(rec.ccdata$genos,c(rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases)),"dominant")
}

