% Generated by roxygen2 (4.0.1): do not edit by hand
\name{esm.p.perm}
\alias{esm.p.perm}
\title{Estimate ESM_K p-value by permutation}
\usage{
esm.p.perm(ccdata, ccstatus, nperms, k)
}
\arguments{
\item{ccdata}{A matrix of markers (columns) and individuals (rows).  Data are coded as the number of copies of the minor allele.}

\item{ccstatus}{A vector of binary phenotype labels.  0 = control, 1 = case.}

\item{nperms}{Number of permutations to perform}

\item{k}{Number of markers to use for ESM_K statistic}
}
\value{
The test statistic, Monte-carlo estimate of P(perm stat >= observed data), and a Z-score based on the permutation distribution.
}
\description{
Estimate ESM_K p-value by permutation
}
\examples{
data(rec.ccdata)
rec.ccdata.status = c( rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases))
#Filter sites: 0 <= MAF in cases < 0.05 && r^2 between pairs < 0.8
keep = filter_sites(rec.ccdata$genos,rec.ccdata.status,0,5e-2,0.8)
rec.ccdata.esm.p = esm.p.perm( rec.ccdata$genos[,which(keep==1)], rec.ccdata.status, 10, 50 )
}
\references{
Thornton, K. R., Foran, A. J., & Long, A. D. (2013). Properties and Modeling of GWAS when Complex Disease Risk Is Due to Non-Complementing, Deleterious Mutations in Genes of Large Effect. PLoS Genetics, 9(2), e1003258. doi:10.1371/journal.pgen.1003258
}

