% Generated by roxygen2 (4.0.2): do not edit by hand
\name{esm_perm_binary}
\alias{esm_perm_binary}
\title{Obtain permutaion distribution of the ESM_K statistic for case/control data}
\usage{
esm_perm_binary(ccdata, ccstatus, nperms, k)
}
\arguments{
\item{ccdata}{A matrix of markers (columns) and individuals (rows).  Data are coded as the number of copies of the minor allele.}

\item{ccstatus}{A vector of binary phenotype labels.  0 = control, 1 = case.}

\item{nperms}{Number of permutations to perform}

\item{k}{Number of markers to use for ESM_K statistic}
}
\value{
A vector of the permuted test statistic values
}
\description{
Obtain permutaion distribution of the ESM_K statistic for case/control data
}
\examples{
data(rec.ccdata)
status = c(rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases))
#filter out common alleles and marker pairs in high LD
keep = filter_sites(rec.ccdata$genos,status,0,0.05,0.8)
rec.ccdata.chisq = chisq_per_marker(rec.ccdata$genos[,which(keep==1)],status)
rec.ccdata.esm = esm( rec.ccdata.chisq, 50 )
rec.ccdata.esm.permdist = esm_perm_binary(rec.ccdata$genos[,which(keep==1)],status,100,50)
}
\references{
Thornton, K. R., Foran, A. J., & Long, A. D. (2013). Properties and Modeling of GWAS when Complex Disease Risk Is Due to Non-Complementing, Deleterious Mutations in Genes of Large Effect. PLoS Genetics, 9(2), e1003258. doi:10.1371/journal.pgen.1003258
}

