% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{esmw}
\alias{esmw}
\title{weighted verstion of Association stat from Thornton, Foran, and Long (2013) PLoS Genetics}
\usage{
esmw(scores, weights, K)
}
\arguments{
\item{scores}{A vector of single-marker association test scores, on a -log10 scale}

\item{weights}{A vector of weights to use for each marker, such as dbetat(MAF,1,25)}

\item{K}{the number of markers used to calculate ESM_K}
}
\value{
The ESM_K test statistic value
}
\description{
weighted verstion of Association stat from Thornton, Foran, and Long (2013) PLoS Genetics
}
\examples{
data(rec.ccdata)
status = c(rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases))
#filter out common alleles and marker pairs in high LD
keep = filter_sites(rec.ccdata$genos,status,0,0.05,0.8)
MAF = colSums(rec.ccdata$genos[,which(keep==1)])/(2*nrow(rec.ccdata$genos))
b.weight = dbeta(MAF,1,25)
lin.weight = rep(1,length(MAF))
rec.ccdata.chisq = chisq_per_marker(rec.ccdata$genos[,which(keep==1)],status)
rec.ccdata.esm = esm( rec.ccdata.chisq,50 )
rec.ccdata.esm.lin =  esmw( rec.ccdata.chisq,lin.weight, 50 )
rec.ccdata.esm.beta = esmw( rec.ccdata.chisq,b.weight, 50 )
}
\references{
Thornton, K. R., Foran, A. J., & Long, A. D. (2013). Properties and Modeling of GWAS when Complex Disease Risk Is Due to Non-Complementing, Deleterious Mutations in Genes of Large Effect. PLoS Genetics, 9(2), e1003258. doi:10.1371/journal.pgen.1003258
}

