% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LLcollapse}
\alias{LLcollapse}
\title{Calculates Li and Leal's collapsed variant statistic, v_c}
\usage{
LLcollapse(ccdata, ccstatus, maf, maf_controls = FALSE)
}
\arguments{
\item{ccdata}{A matrix of markers (columns) and individuals (rows).  Data are coded as the number of copies of the minor allele.}

\item{ccstatus}{A vector of binary phenotype labels.  0 = control, 1 = case.}

\item{maf}{Only consider variants whose minor allele frequencies are <= maf}

\item{maf_controls}{If true, calculate mafs from controls only.  Otherwise, use all individuals}
}
\value{
A chi-squared statistic based on a 2x2 table of the number of cases and controls with and without rare alleles.  Yate's correction is applied.
}
\description{
Calculates Li and Leal's collapsed variant statistic, v_c
}
\examples{
data(rec.ccdata)
status = c(rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases))
LL = LLcollapse(rec.ccdata$genos,status,0.01)
}
\references{
Li, B., & Leal, S. (2008). Methods for detecting associations with rare variants for common diseases: application to analysis of sequence data. The American Journal of Human Genetics, 83(3), 311-321.
}

