% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LLcollapse_perm}
\alias{LLcollapse_perm}
\title{Permutation distribution of Li and Leal's collapsed variant statistic, v_c}
\usage{
LLcollapse_perm(ccdata, ccstatus, nperms, maf, maf_controls = FALSE)
}
\arguments{
\item{ccdata}{A matrix of markers (columns) and individuals (rows).  Data are coded as the number of copies of the minor allele.}

\item{ccstatus}{A vector of binary phenotype labels.  0 = control, 1 = case.}

\item{nperms}{The number of permutations to perform}

\item{maf}{Only consider variants whose minor allele frequencies are <= maf}

\item{maf_controls}{If true, calculate mafs from controls only.  Otherwise, use all individuals}
}
\value{
The non-centrality parameter of a chi-squared distribution.  This is obtained using the proportion of controls and cases with rare variants.
}
\description{
Permutation distribution of Li and Leal's collapsed variant statistic, v_c
}
\examples{
data(rec.ccdata)
status = c(rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases))
LL.perm = LLcollapse_perm(rec.ccdata$genos,status,10,0.01)
}
\references{
Li, B., & Leal, S. (2008). Methods for detecting associations with rare variants for common diseases: application to analysis of sequence data. The American Journal of Human Genetics, 83(3), 311-321.
}

