% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logitTests.R
\name{logit.additive}
\alias{logit.additive}
\title{Association testing under an additive model}
\usage{
logit.additive(x, status)
}
\arguments{
\item{x}{An array of genotypes, coded as number of copies of minor allele}

\item{status}{An array of 0 = control, 1 = case.  length(status) must equal length(x)}
}
\value{
The p-value of a logistic regression of case/control status onto genotype.
}
\description{
Association testing under an additive model
}

