% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logitTests.R
\name{logit.dominant}
\alias{logit.dominant}
\title{Association testing under a dominant model}
\usage{
logit.dominant(x, status)
}
\arguments{
\item{x}{An array of genotypes, coded as number of copies of minor allele}

\item{status}{An array of 0 = control, 1 = case.  length(status) must equal length(x)}
}
\value{
The p-value of a logistic regression of case/control status onto genotype.
}
\description{
Association testing under a dominant model
}
\note{
The test works by recoding 2 genotypes as 1 (e.g., one copy of minor allele is the same as two)
}

