% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logitTests.R
\name{logit.recessive}
\alias{logit.recessive}
\title{Association testing under a recessive model}
\usage{
logit.recessive(x, status)
}
\arguments{
\item{x}{An array of genotypes, coded as number of copies of minor allele}

\item{status}{An array of 0 = control, 1 = case.  length(status) must equal length(x)}
}
\value{
The p-value of a logistic regression of case/control status onto genotype.
}
\description{
Association testing under a recessive model
}
\note{
The test works by recoding 1 genotypes as 0 (e.g., one copy of minor allele is the same as none)
}

