% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/perms.R
\name{MB.p.perm}
\alias{MB.p.perm}
\title{Estimate Madsen-Browning p-value by permutation}
\usage{
MB.p.perm(ccdata, ccstatus, nperms)
}
\arguments{
\item{ccdata}{A matrix of markers (columns) and individuals (rows).  Data are coded as the number of copies of the minor allele.}

\item{ccstatus}{A vector of binary phenotype labels.  0 = control, 1 = case.}

\item{nperms}{Number of permutations to perform}
}
\value{
A list of statistics, p-values, and Z-scores, one for each of the three models
}
\description{
Estimate Madsen-Browning p-value by permutation
}
\examples{
data(rec.ccdata)
rec.ccdata.status = c( rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases))
#Filter sites: 0 <= MAF in cases < 0.05 && r^2 between pairs < 0.8
keep = filter_sites(rec.ccdata$genos,rec.ccdata.status,0,5e-2,0.8)
rec.ccdata.MB.p = MB.p.perm( rec.ccdata$genos[,which(keep==1)], rec.ccdata.status, 10 )
}
\references{
Madsen, B. E., & Browning, S. R. (2009). A groupwise association test for rare mutations using a weighted sum statistic. PLoS Genetics, 5(2), e1000384. doi:10.1371/journal.pgen.1000384
}

