% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MB_perm}
\alias{MB_perm}
\title{Get permutation distribution of Madsen-Browning test statistics}
\usage{
MB_perm(ccdata, ccstatus, nperms)
}
\description{
Get permutation distribution of Madsen-Browning test statistics
}
\examples{
data(rec.ccdata)
status = c(rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases))
#filter out common alleles and marker pairs in high LD
keep = filter_sites(rec.ccdata$genos,status,0,0.05,0.8)
mbstats = MBstat( rec.ccdata$genos[,which(keep==1)], status )
mbstats.perm = MB_perm( rec.ccdata$genos[,which(keep==1)], status, 10000 )
}

