% Generated by roxygen2 (4.0.2): do not edit by hand
\name{norm.genos}
\alias{norm.genos}
\title{Turn genotypes into normalized genotypes}
\usage{
norm.genos(genotypes)
}
\arguments{
\item{genotypes}{An individuals x mutation matrix, coded 0,1,2 = number of copies of minor alleles}
}
\description{
Turn genotypes into normalized genotypes
}
\note{
copied from Golan's original code at https://sites.google.com/site/davidgolanshomepage/software/pcgc
}

