% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ProductMoment}
\alias{ProductMoment}
\title{Pearson's product-moment correlation}
\usage{
ProductMoment(x, y)
}
\arguments{
\item{x}{A vector of values.}

\item{y}{A vector of values.}
}
\value{
The correlation coefficient between x and y. This should/will be equal to R's cor(x,y).
}
\description{
Pearson's product-moment correlation
}
\note{
This implementation is based on a copy of [libsequence's](http://github.com/molpopgen/libsequence) template function object Sequence::ProductMoment in <Sequence/Correlations.hpp>
}

