% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{rec.ccdata}
\alias{rec.ccdata}
\title{Example case/control data}
\format{A simulated case/control panel with 3,000 controls plus 3,000 cases.  The data are a list composed of the following elements:
genos = the genotypes.  rows = individuals, columns = markers.  Values are number of copies of the derived allele.\cr
pos = mutation positions.\cr
ncontrols,ncases = number of controls and cases, respectively.  The controls come first in genos.\cr
burdens = A matrix.  Each row is the number of causative mutations inherited from each parent.\cr
phenos = A matrix representing phenotype values for each individual.  The first column is the genetic component.  The second is the random component.\cr
neutral = An integer.  The first neutral columns in genos are markers with an effect size of 0.\cr
causative = An integer.  The remaining causative columns following the neutral ones have nonzero effect sizes.\cr}
\usage{
rec.ccdata
}
\description{
Example case/control data
}
\details{
The data were simulated according the the model described in http://www.ncbi.nlm.nih.gov/pubmed/23437004, using the parameters
from that paper.  The mean effect size was 0.1.
}
\keyword{datasets}

