% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{rec.esizes}
\alias{rec.esizes}
\title{Effect sizes for example data}
\format{The effect size block corresponding to rec.ccdata.  The format is a matrix of 4 columns, and there is one row for each mutation in the entire population. (In other words, nrow(rec.esize) > ncol(rec.ccdata$genos)).  The columns are:\cr
mutation positions\cr
mutation effect sizes\cr
Number of occurrences of each mutation in the population.  The population size was 2e4 diploids, meaning that the mutation frequency is the value in the matrix divided by 4e4\cr
The age of the mutation, in generations.}
\usage{
rec.esizes
}
\description{
Effect sizes for example data
}
\keyword{datasets}

