% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc.Gij}
\alias{calc.Gij}
\title{Equation 5 from www.pnas.org/cgi/doi/10.1073/pnas.1419064111}
\usage{
calc.Gij(genotypes)
}
\arguments{
\item{genotypes}{A matrix of genotypes}
}
\description{
Equation 5 from www.pnas.org/cgi/doi/10.1073/pnas.1419064111
}
\note{
copied from Golan's original code at https://sites.google.com/site/davidgolanshomepage/software/pcgc,
}
\references{
Golan et al. (2014) Measuring missing heritability: Inferring the contribution of common variants. www.pnas.org/cgi/doi/10.1073/pnas.1419064111
}

