% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cAlpha}
\alias{cAlpha}
\title{The c-alpha statistic}
\usage{
cAlpha(data, status)
}
\arguments{
\item{data}{A matrix of markers (columns) and individuals (rows).  Data are coded as the number of copies of the minor allele.}

\item{status}{A vector of binary phenotype labels.  0 = control, 1 = case.}
}
\description{
The c-alpha statistic
}
\examples{
data(rec.ccdata)
status = c( rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases) )
rec.ccdata.calpha = cAlpha(rec.ccdata$genos,status)
}

