% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ccpcgc.R
\name{cc2pcgc}
\alias{cc2pcgc}
\title{Write a case/control block to files that will be used
as the input to the pcgc software from https://sites.google.com/site/davidgolanshomepage/software/pcgc,
which implements the methods in www.pnas.org/cgi/doi/10.1073/pnas.1419064111}
\usage{
cc2pcgc(genotypes, status, filenamebase)
}
\arguments{
\item{genotypes}{A matrix of genotypes, coded as number of copies of minor allele}

\item{status}{An array of 0 = control, 1 = case.  length(status) must equal nrow(x)}

\item{filenamebase}{The prefix to use for output file names.  See note below}
}
\description{
Write a case/control block to files that will be used
as the input to the pcgc software from https://sites.google.com/site/davidgolanshomepage/software/pcgc,
which implements the methods in www.pnas.org/cgi/doi/10.1073/pnas.1419064111
}
\examples{
\dontrun{
data(rec.ccdata)
rec.ccdata.status = c( rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases))
cc2pcgc(rec.ccdata$genos,rec.ccdata.status,"rec.ccdata.pcgc")
}
}
\references{
Golan et al. (2014) Measuring missing heritability: Inferring the contribution of common variants. www.pnas.org/cgi/doi/10.1073/pnas.1419064111
}

