% Generated by roxygen2 (4.0.1): do not edit by hand
\name{esm}
\alias{esm}
\title{Association stat from Thornton, Foran, and Long (2013) PLoS Genetics}
\usage{
esm(scores, K)
}
\arguments{
\item{scores}{A vector of single-marker association test scores, on a -log10 scale}

\item{K}{the number of markers used to calculate ESM_K}
}
\value{
The ESM_K test statistic value
}
\description{
Association stat from Thornton, Foran, and Long (2013) PLoS Genetics
}
\note{
See http://www.ncbi.nlm.nih.gov/pubmed/23437004 for detail on the test statistic
}

