% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MB_perm}
\alias{MB_perm}
\title{Get permutation distribution of Madsen-Browning test statistics}
\usage{
MB_perm(ccdata, ccstatus, nperms)
}
\arguments{
\item{ccdata}{A matrix of markers (columns) and individuals (rows).  Data are coded as the number of copies of the minor allele.}

\item{ccstatus}{A vector of binary phenotype labels.  0 = control, 1 = case.}

\item{nperms}{The number of permutations to perform}
}
\value{
A data frame of permuted statistics
}
\description{
Get permutation distribution of Madsen-Browning test statistics
}
\examples{
data(rec.ccdata)
status = c(rep(0,rec.ccdata$ncontrols),rep(1,rec.ccdata$ncases))
#filter out common alleles and marker pairs in high LD
keep = filter_sites(rec.ccdata$genos,status,0,0.05,0.8)
mbstats = MBstat( rec.ccdata$genos[,which(keep==1)], status )
mbstats.perm = MB_perm( rec.ccdata$genos[,which(keep==1)], status, 100 )
}
\references{
Madsen, B. E., & Browning, S. R. (2009). A groupwise association test for rare mutations using a weighted sum statistic. PLoS Genetics, 5(2), e1000384. doi:10.1371/journal.pgen.1000384
}

