% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MBweights}
\alias{MBweights}
\title{Calculate Madsen-Browning weights.}
\usage{
MBweights(ccdata, ccstatus)
}
\arguments{
\item{ccdata}{A matrix of markers (columns) and individuals (rows).  Data are coded as the number of copies of the minor allele.}

\item{ccstatus}{A vector of binary phenotype labels.  0 = control, 1 = case.}
}
\value{
An array of weights, one for each column in data.
}
\description{
Calculate Madsen-Browning weights.
}
\details{
Calculation is done under the "general genetic model" defined in Madsen and Browning.
}
\references{
Madsen, B. E., & Browning, S. R. (2009). A groupwise association test for rare mutations using a weighted sum statistic. PLoS Genetics, 5(2), e1000384. doi:10.1371/journal.pgen.1000384
}

