% Generated by roxygen2 (4.0.1): do not edit by hand
\name{norm.genos}
\alias{norm.genos}
\title{Turn genotypes into normalized genotypes}
\usage{
norm.genos(genotypes)
}
\description{
Turn genotypes into normalized genotypes
}
\note{
copied from Golan's original code at https://sites.google.com/site/davidgolanshomepage/software/pcgc,
}

